fixed array unittest
the bug occurs when reading a string inside an array. it is because TokenSrc cached a token but String::operator>>() try to use TokenSrc::Stream() to read chacters directly. It will then skip the cached tokens and get wrong result.

the fix is to add an GetChar() to TokenSrc that will grab a char from its cache, if any, before reading from the Stream().